requirements.txt
attrs==22.1.0
autopep8==2.0.0
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.0
tomli==2.0.1
pg8000>=1.30.3
python-dotenv==1.0.0